Ability//Move handling
Added handling to display abilities and moves in the load script.
Also added a lookup button to both to automatically reference names to the 'dexes and fill other data from there.
both have a dynamic number of entries
also changed systems to allow for multiple windows to be active simultaneously